Add files via upload
I am still working on some of the files and stuff. But I am tired so I am gonna head to bed. I am in the process of doing the Menu Page and the Information Page. I am also trying to add Back Buttons to all the other pages so people can go back if they need too. Added some pretty colors. Gonna make the labels bold eventually so it is easier to see with the red background. 

Asim: Look at my InfoGUI file. I am trying to mess with the panels and get the back button to be at the button of the page. After 1-2 hours of messing with it, I am giving up until tomorrow. I suck with these panels

OK. Night!!

I am using netbeans btw...